refactor: rename repositories and memoize mongo client
