+++ lib/HinataCronJob.js
@@ -1,7 +1,7 @@
 import { CronJob } from 'cron';
 import { EventEmitter } from 'events';
 import { getLastWeekStatsForGroup } from './api';
-import HinataRepository from './repositories';
+import HinataRepository from './repository';
 import { createEventHandler, deepEquals } from './shared';
 
 const FEMALE_3DIV_GROUPA_ID = '3147';
