+++ lib/index.js
@@ -1,12 +1,12 @@
 import 'dotenv/config';
 import VError from 'verror';
 import HinataCronJob from './HinataCronJob';
-import HinataRepository from './repositories';
+import HinataRepository from './repository';
 import { customLog } from './shared';
 import { HinataTelegramBot, TelegramBotUtils } from './telegramBot';
 
-async function main() {
-  const hinataRepository = await HinataRepository();
+function main() {
+  const hinataRepository = HinataRepository();
   const hinataCronJob = HinataCronJob({ repository: hinataRepository });
   const hinataTelegramBot = HinataTelegramBot({ repository: hinataRepository });
   const botUtils = TelegramBotUtils({ repository: hinataRepository });
