--- lib/repositories/index.js
@@ -1,7 +1,0 @@
-import MongoRepository from './mongo';
-
-export default async function HinataRepository() {
-  const hinataRepository = await MongoRepository();
-
-  return hinataRepository;
-}
