+++ lib/repository/index.js
@@ -1,0 +1,7 @@
+import MongoRepository from './mongo';
+
+export default function HinataRepository() {
+  const hinataRepository = MongoRepository();
+
+  return hinataRepository;
+}
