+++ lib/shared.js
@@ -18,9 +18,25 @@ function isEmpty(obj) {
   return Object.keys(obj).length === 0;
 }
 
+function memoizeOne(memoFn) {
+  let result;
+  let isCalled = false;
+
+  const memoized = () => {
+    if (!isCalled) {
+      result = memoFn();
+      isCalled = true;
+    }
+    return result;
+  };
+
+  return memoized;
+}
+
 export {
   createEventHandler,
   customLog,
   deepEquals,
   isEmpty,
+  memoizeOne,
 };
