+++ lib/telegramBot/HinataTelegramBot.js
@@ -1,7 +1,7 @@
 import { EventEmitter } from 'events';
 import TelegramBot from 'node-telegram-bot-api';
 import { lastWeekStatsToText } from '../formatUtils';
-import HinataRepository from '../repositories';
+import HinataRepository from '../repository';
 import { createEventHandler } from '../shared';
 
 const EVENT_CHECK_CRON_JOB_RUNNING = 'check_cron_job_running';
