+++ lib/telegramBot/utils.js
@@ -1,6 +1,6 @@
 import VError from 'verror';
 import { lastWeekStatsToText } from '../formatUtils';
-import HinataRepository from '../repositories';
+import HinataRepository from '../repository';
 import { customLog, isEmpty } from '../shared';
 
 export default function TelegramBotUtils({
